refactor: rename payment-request and payment-request-creation-data
