+++ app/screens/receive-bitcoin-screen/payment/payment-request.ts
@@ -1,99 +1,227 @@
 import { WalletCurrency } from "@app/graphql/generated"
-import { WalletDescriptor } from "@app/types/wallets"
-import { MoneyAmount, WalletAmount, WalletOrDisplayCurrency } from "@app/types/amounts"
-
 import {
-  InvoiceType,
+  CreatePaymentRequestParams,
+  GetFullUriFn,
+  Invoice,
   PaymentRequest,
-  BaseCreatePaymentRequestParams,
-  ConvertMoneyAmountFn,
+  PaymentRequestState,
+  PaymentRequestStateType,
+  PaymentRequestInformation,
 } from "./index.types"
+import { decodeInvoiceString, Network as NetworkLibGaloy } from "@galoymoney/client"
+import { BtcMoneyAmount } from "@app/types/amounts"
+import { getPaymentRequestFullUri } from "./helpers"
 
-export const createPaymentRequest = <T extends WalletCurrency>(
-  params: BaseCreatePaymentRequestParams<T>,
-): PaymentRequest<T> => {
-  // These sets are always available
-  const setType = (type: InvoiceType) => createPaymentRequest({ ...params, type })
-  const setDefaultWalletDescriptor = (defaultWalletDescriptor: WalletDescriptor<T>) =>
-    createPaymentRequest({ ...params, defaultWalletDescriptor })
-  const setConvertMoneyAmount = (convertMoneyAmount: ConvertMoneyAmountFn) =>
-    createPaymentRequest({ ...params, convertMoneyAmount })
+export const createPaymentRequest = (
+  params: CreatePaymentRequestParams,
+): PaymentRequest => {
+  let { state, info } = params
+  if (!state) state = PaymentRequestState.Idle
 
-  const { type, defaultWalletDescriptor, convertMoneyAmount, memo } = params
-
-  // Permissions for the specified type
-  const permissions = {
-    canSetReceivingWalletDescriptor: false,
-    canSetMemo: false,
-    canSetAmount: false,
-  }
-  switch (type) {
-    case "Lightning":
-      permissions.canSetReceivingWalletDescriptor = true
-    case "OnChain":
-      permissions.canSetMemo = true
-      permissions.canSetAmount = true
+  const setState = (state: PaymentRequestStateType) => {
+    if (state === PaymentRequestState.Loading)
+      return createPaymentRequest({ ...params, state, info: undefined })
+    return createPaymentRequest({ ...params, state })
   }
 
-  // Permission based sets
-  let setReceivingWalletDescriptor:
-    | ((receivingWalletDescriptor: WalletDescriptor<T>) => PaymentRequest<T>)
-    | undefined = undefined
-  if (permissions.canSetReceivingWalletDescriptor) {
-    setReceivingWalletDescriptor = (receivingWalletDescriptor) =>
-      createPaymentRequest({ ...params, receivingWalletDescriptor })
-  }
-  let setMemo: ((memo: string) => PaymentRequest<T>) | undefined = undefined
-  if (permissions.canSetMemo) {
-    setMemo = (memo) => createPaymentRequest({ ...params, memo })
-  }
-  let setAmount:
-    | ((unitOfAccountAmount: MoneyAmount<WalletOrDisplayCurrency>) => PaymentRequest<T>)
-    | undefined = undefined
-  if (permissions.canSetAmount) {
-    setAmount = (unitOfAccountAmount) =>
-      createPaymentRequest({ ...params, unitOfAccountAmount })
+  // The hook should setState(Loading) before calling this
+  const generateQuote: () => Promise<PaymentRequest> = async () => {
+    const { creationData: pr, mutations } = params
+
+    let info: PaymentRequestInformation | undefined
+
+    // On Chain BTC
+    if (pr.type === Invoice.OnChain) {
+      const { data, errors } = await mutations.onChainAddressCurrent({
+        variables: { input: { walletId: pr.receivingWalletDescriptor.id } },
+      })
+
+      if (pr.settlementAmount && pr.settlementAmount.currency !== WalletCurrency.Btc)
+        throw new Error("On-chain invoices only support BTC")
+
+      const address = data?.onChainAddressCurrent?.address || undefined
+
+      const getFullUriFn: GetFullUriFn = ({ uppercase, prefix }) =>
+        getPaymentRequestFullUri({
+          type: Invoice.OnChain,
+          input: address || "",
+          amount: pr.settlementAmount?.amount,
+          memo: pr.memo,
+          uppercase,
+          prefix,
+        })
+
+      info = {
+        data: address
+          ? {
+              invoiceType: Invoice.OnChain,
+              getFullUriFn,
+              address,
+              amount: pr.settlementAmount as BtcMoneyAmount,
+              memo: pr.memo,
+            }
+          : undefined,
+        applicationErrors: data?.onChainAddressCurrent?.errors,
+        gqlErrors: errors,
+      }
+
+      // Lightning without Amount
+    } else if (pr.type === Invoice.Lightning && pr.unitOfAccountAmount === undefined) {
+      const { data, errors } = await mutations.lnNoAmountInvoiceCreate({
+        variables: {
+          input: {
+            walletId: pr.receivingWalletDescriptor.id,
+            memo: pr.memo,
+          },
+        },
+      })
+
+      const dateString = decodeInvoiceString(
+        data?.lnNoAmountInvoiceCreate.invoice?.paymentRequest ?? "",
+        pr.network as NetworkLibGaloy,
+      ).timeExpireDateString
+
+      const getFullUriFn: GetFullUriFn = ({ uppercase, prefix }) =>
+        getPaymentRequestFullUri({
+          type: Invoice.Lightning,
+          input: data?.lnNoAmountInvoiceCreate.invoice?.paymentRequest || "",
+          amount: pr.settlementAmount?.amount,
+          memo: pr.memo,
+          uppercase,
+          prefix,
+        })
+
+      info = {
+        data: data?.lnNoAmountInvoiceCreate.invoice
+          ? {
+              invoiceType: Invoice.Lightning,
+              ...data?.lnNoAmountInvoiceCreate.invoice,
+              expiresAt: dateString ? new Date(dateString) : undefined,
+              getFullUriFn,
+            }
+          : undefined,
+        applicationErrors: data?.lnNoAmountInvoiceCreate?.errors,
+        gqlErrors: errors,
+      }
+
+      // Lightning with BTC Amount
+    } else if (
+      pr.type === Invoice.Lightning &&
+      pr.unitOfAccountAmount?.currency === WalletCurrency.Btc &&
+      pr.settlementAmount?.amount
+    ) {
+      const { data, errors } = await mutations.lnInvoiceCreate({
+        variables: {
+          input: {
+            walletId: pr.receivingWalletDescriptor.id,
+            amount: pr.settlementAmount?.amount,
+            memo: pr.memo,
+          },
+        },
+      })
+
+      const dateString = decodeInvoiceString(
+        data?.lnInvoiceCreate.invoice?.paymentRequest ?? "",
+        pr.network as NetworkLibGaloy,
+      ).timeExpireDateString
+
+      const getFullUriFn: GetFullUriFn = ({ uppercase, prefix }) =>
+        getPaymentRequestFullUri({
+          type: Invoice.Lightning,
+          input: data?.lnInvoiceCreate.invoice?.paymentRequest || "",
+          amount: pr.settlementAmount?.amount,
+          memo: pr.memo,
+          uppercase,
+          prefix,
+        })
+
+      info = {
+        data: data?.lnInvoiceCreate.invoice
+          ? {
+              invoiceType: Invoice.Lightning,
+              ...data?.lnInvoiceCreate.invoice,
+              expiresAt: dateString ? new Date(dateString) : undefined,
+              getFullUriFn,
+            }
+          : undefined,
+        applicationErrors: data?.lnInvoiceCreate?.errors,
+        gqlErrors: errors,
+      }
+      // Lightning with USD Amount
+    } else if (
+      pr.type === Invoice.Lightning &&
+      pr.unitOfAccountAmount?.currency === WalletCurrency.Usd &&
+      pr.settlementAmount?.amount
+    ) {
+      const { data, errors } = await mutations.lnUsdInvoiceCreate({
+        variables: {
+          input: {
+            walletId: pr.receivingWalletDescriptor.id,
+            amount: pr.settlementAmount?.amount,
+            memo: pr.memo,
+          },
+        },
+      })
+
+      const dateString = decodeInvoiceString(
+        data?.lnUsdInvoiceCreate.invoice?.paymentRequest ?? "",
+        pr.network as NetworkLibGaloy,
+      ).timeExpireDateString
+
+      const getFullUriFn: GetFullUriFn = ({ uppercase, prefix }) =>
+        getPaymentRequestFullUri({
+          type: Invoice.Lightning,
+          input: data?.lnUsdInvoiceCreate.invoice?.paymentRequest || "",
+          amount: pr.settlementAmount?.amount,
+          memo: pr.memo,
+          uppercase,
+          prefix,
+        })
+
+      info = {
+        data: data?.lnUsdInvoiceCreate.invoice
+          ? {
+              invoiceType: Invoice.Lightning,
+              ...data?.lnUsdInvoiceCreate.invoice,
+              expiresAt: dateString ? new Date(dateString) : undefined,
+              getFullUriFn,
+            }
+          : undefined,
+        applicationErrors: data?.lnUsdInvoiceCreate?.errors,
+        gqlErrors: errors,
+      }
+
+      // Paycode
+    } else if (pr.type === Invoice.PayCode && pr.username) {
+      const getFullUriFn: GetFullUriFn = ({ uppercase, prefix }) =>
+        getPaymentRequestFullUri({
+          type: Invoice.PayCode,
+          input: pr.username || "",
+          uppercase,
+          prefix,
+        })
+
+      info = {
+        data: {
+          invoiceType: Invoice.PayCode,
+          username: pr.username,
+          getFullUriFn,
+        },
+        applicationErrors: undefined,
+        gqlErrors: undefined,
+      }
+    } else {
+      info = undefined
+      throw new Error("Unknown Payment Request Type Encountered - Please Report")
+    }
+
+    let state: PaymentRequestStateType = PaymentRequestState.Created
+    if (!info || info.applicationErrors?.length || info.gqlErrors?.length || !info.data) {
+      state = PaymentRequestState.Error
+    }
+
+    return createPaymentRequest({ ...params, info, state })
   }
 
-  // Set default receiving wallet descriptor
-  let { receivingWalletDescriptor } = params
-  if (!receivingWalletDescriptor) {
-    receivingWalletDescriptor = defaultWalletDescriptor
-  }
-
-  // OnChain only on BTC
-  if (type === "OnChain") {
-    receivingWalletDescriptor = params.bitcoinWalletDescriptor as WalletDescriptor<T>
-  }
-
-  // Set settlement amount if unit of account amount is set
-  let { unitOfAccountAmount } = params
-  let settlementAmount: WalletAmount<T> | undefined = undefined
-  if (unitOfAccountAmount) {
-    settlementAmount = convertMoneyAmount(
-      unitOfAccountAmount,
-      receivingWalletDescriptor.currency,
-    )
-  }
-
-  return {
-    ...params,
-    ...permissions,
-    setType,
-    setDefaultWalletDescriptor,
-    setConvertMoneyAmount,
-    receivingWalletDescriptor,
-
-    // optional sets
-    setReceivingWalletDescriptor,
-    setMemo,
-    setAmount,
-
-    // optional data
-    unitOfAccountAmount,
-    settlementAmount,
-    memo,
-
-    canUsePaycode: Boolean(params.username),
-  }
+  return { ...params, state, info, generateRequest: generateQuote, setState }
 }
